op3: Set eBPF support level
Right now, the kernel in use does not have complete eBPF support,
which is now required in Android 12. For now, declare the device
as eBPF-incapable.

All changes in topic:twelve-ultralegacy-devices on the LineageOS
Gerrit are required for this property to work. Otherwise, the
device will not boot.

A more permanent fix would be to port Linux kernel 4.9 to this
device, since it has full eBPF support. That way, we can remove
this hack later.